pyproject.toml
[project]
name = "llmops_databricks"
dynamic = ["version"]
description = "llmops-databricks-course-yilmazsal - LLMOps Course on Databricks"
requires-python = ">=3.12, <3.13"
dependencies = [
    "cffi==1.17.1",
     "cloudpickle==3.1.1",
    "numpy==2.4.3",
    "pandas==2.2.3",
    "pyarrow==23.0.1",
    "databricks-sdk==0.102.0",
    "pydantic==2.12.5",
    "loguru==0.7.3",
    "python-dotenv==1.2.2",
    "databricks-vectorsearch==0.66",
    "openai==2.29.0",
    "databricks-mcp==0.9.0",
    "backoff==2.2.1",
    "mlflow==3.10.1",
    "nest-asyncio==1.6.0",
    "arxiv==2.4.1",
    "databricks-agents==1.9.3",
    "psycopg==3.3.3",
    "psycopg-pool==3.3.0",
    "psycopg[binary]==3.3.3",
    "azure-ai-documentintelligence==1.0.2"
]

[project.optional-dependencies]
dev = ["databricks-connect>=18.1.2, <19",
       "ipykernel>=7.2.0, <8",
       "pip>=25.0.1, <26",
       "pre-commit>=4.5.1, <5",
       "claudetracing>=0.2.0, <1",]

ci = ["pre-commit>=4.5.1, <5",
      "pytest>=9.0.2, <10"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"your_custom_package" = ["../../../project_config.yml"]

[build-system]
requires = ["setuptools>=72.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["llmops_databricks*"]

[tool.setuptools.dynamic]
version = { file = "version.txt" }

[tool.pytest.ini_options]
addopts = "-s --no-header --no-summary"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 90

select = [
    "F",    # pyflakes rules
    "E",    # pycodestyle error rules
    "W",    # pycodestyle warning rules
    "B",    # flake8-bugbear rules
    "I",    # isort rules
    "UP",   # pyupgrade rules
    "SIM",  # flake8-simplify rules
    "ERA",  # eradicate rules
    "C",    # pylint convention rules
    "ANN"   # flake8-annotations rules
]

ignore = [
    "D107",   # Missing docstring in `__init__`
    "D203",   # incorrect-blank-line-before-class
    "D213",   # multi-line-summary-second-line
    "C901",   # too complex
    "ERA001", # commented out code
    "ANN204",
    "ANN102",
    "ANN101"
]

[tool.ruff.format]
indent-style = "space"
